requirements.txt
python-telegram-bot~=12.5.1
textblob~=0.15.3
emoji~=0.5.4
Chatterbot~=1.0.5
google-api-python-client~=1.7.11
google-auth~=1.11.0
google-auth-httplib2~=0.0.3
google-auth-oauthlib~=0.4.1
httplib2~=0.17.0
